refactor(runtime-api): extract session handlers
